chore: Merge `Conversations` with `MessageStore`
